Add network url to next dev
