+++ packages/next/cli/next-start.ts
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 
-import os from 'os'
 import arg from 'next/dist/compiled/arg/index.js'
 import { startServer } from '../server/lib/start-server'
 import { getPort, printAndExit } from '../server/lib/utils'
@@ -8,21 +7,7 @@ import * as Log from '../build/output/log'
 import isError from '../lib/is-error'
 import { getProjectDir } from '../lib/get-project-dir'
 import { cliCommand } from '../lib/commands'
-
-function networkHosts() {
-  const interfaces = os.networkInterfaces()
-  const hosts: string[] = []
-
-  Object.keys(interfaces).forEach((key) =>
-    interfaces[key]
-      ?.filter(({ family, address }) => {
-        return family === 'IPv4' && !address.includes('127.0.0.1')
-      })
-      .forEach(({ address }) => hosts.push(address))
-  )
-
-  return hosts
-}
+import { getNetworkHost } from '../lib/get-network-hosts'
 
 const nextStart: cliCommand = (argv) => {
   const validArgs: arg.Spec = {
@@ -96,15 +81,11 @@ const nextStart: cliCommand = (argv) => {
   })
     .then(async (app) => {
       const appUrl = `http://${app.hostname}:${app.port}`
-
-      const networkUrl = networkHosts()
-        .map((networkHost) => `http://${networkHost}:${app.port}`)
-        .join(' ')
+      const networkUrl = `http://${getNetworkHost()}:${app.port}`
 
       Log.ready(
         `started server on ${host}:${app.port}, url: ${appUrl} (${networkUrl})`
       )
-
       await app.prepare()
     })
     .catch((err) => {
